Various fixes and impreovements based on refactor.
